--- scripts/egp-tbd/createProposal.sh
@@ -1,9 +1,0 @@
-
-# Script to update grants with new values
-
-NETWORK=hardhat
-
-# preview the proposal
-npx hardhat run scripts/egp-tbd/previewProposal.ts --no-compile --network $NETWORK
-
-npx hardhat run scripts/egp-tbd/createProposal.ts --no-compile --network $NETWORK
