+++ scripts/egp27/createProposal.ts
@@ -44,7 +44,7 @@ export async function createUpgradeGrantsProposal() {
 
   console.log("creating the proposal");
 
-  const rawdata = fs.readFileSync("scripts/egp-tbd/proposalArgs.json");
+  const rawdata = fs.readFileSync("scripts/egp27/proposalArgs.json");
   const args: ProposalArgs = JSON.parse(rawdata.toString());
   const {
     targets,
@@ -104,7 +104,7 @@ export async function createUpgradeGrantsProposal() {
 
   const proposalInfo: ProposalInfo = Object.fromEntries(proposalArgs);
   const data = JSON.stringify(proposalInfo, null, 2);
-  fs.writeFileSync("scripts/egp-tbd/proposalInfo.json", data);
+  fs.writeFileSync("scripts/egp27/proposalInfo.json", data);
 }
 
 // We recommend this pattern to be able to use async/await everywhere
