+++ scripts/egp27/executeProposal.sh
@@ -4,4 +4,4 @@
 NETWORK=hardhat
 
 # executing the proposal
-npx hardhat run scripts/egp-tbd/executeProposal.ts --no-compile --network $NETWORK
+npx hardhat run scripts/egp27/executeProposal.ts --no-compile --network $NETWORK
