+++ scripts/egp27/executeProposal.ts
@@ -26,7 +26,7 @@ export async function main() {
   const coreVotingContract = CoreVoting__factory.connect(coreVoting, signer);
   const timelockContract = Timelock__factory.connect(timeLock, signer);
 
-  const rawdata = fs.readFileSync("scripts/egp-tbd/proposalInfo.json");
+  const rawdata = fs.readFileSync("scripts/egp27/proposalInfo.json");
   const proposalInfo: ProposalInfo = JSON.parse(rawdata.toString());
   const { proposalId, targets, callDatas, targetsTimeLock, calldatasTimeLock } =
     proposalInfo;
@@ -53,7 +53,7 @@ export async function main() {
     console.log("err", err.reason);
   }
 
-  const granteeAddresses = grants.grantUpdatesForEGPTBD.map((g) => g.who);
+  const granteeAddresses = grants.grantUpdatesForEGP27.map((g) => g.who);
 
   const grantsAfterProposal = await fetchGrantsByAddress(vestingVault, signer);
   console.log("logging all grants");
