+++ scripts/egp27/proposalInfo.json
